fastAPI/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
requests
chromadb
feedparser
xmltodict
rapidfuzz
google-genai
spacy
